(PC-29540)[API] refactor: move `domain.allocine` to `core.providers.allocine`
