+++ api/src/pcapi/core/providers/commands.py
@@ -4,7 +4,7 @@ from time import time
 import click
 import sqlalchemy as sa
 
-from pcapi.core.providers import allocine_movie_list
+from pcapi.core.providers import allocine
 import pcapi.core.providers.repository as providers_repository
 from pcapi.local_providers import provider_manager
 from pcapi.utils.blueprint import Blueprint
@@ -20,7 +20,7 @@ logger = logging.getLogger(__name__)
 
 @blueprint.cli.command("synchronize_allocine_products")
 def synchronize_allocine_products() -> None:
-    allocine_movie_list.synchronize_products()
+    allocine.synchronize_products()
 
 
 @blueprint.cli.command("synchronize_venue_providers_apis")
